% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.R
\name{analyze}
\alias{analyze}
\title{Plots the average, min and max inflammation over time.}
\usage{
analyze(filename, output = NULL)
}
\arguments{
\item{filename}{character string of a csv file}
}
\value{
plots of avg, min and max inflammation
}
\description{
This function takes a csv file of patients inflammation and outputs average,
min and max inflammation for each day.
}
\examples{
analyze("inflammation-01.csv")
}
