% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_all.R
\name{analyze_all}
\alias{analyze_all}
\title{Produces PDF files with avg, max and min of inflammation over all the days.}
\usage{
analyze_all(pattern)
}
\arguments{
\item{pattern}{pattern of the name of the files}
}
\value{
PDF files with graphs
}
\description{
This function takes csv files of patients inflammation and prints pdf files with average,
min and max inflammation for each day.
}
\examples{
analyze_all("inflammation*.csv")
}
