% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_df.r, R/summarize.r
\docType{data}
\name{example_df}
\alias{example_df}
\title{Example data set}
\format{
\subsection{\code{example_df}}{

A data frame with 100 rows and 3 columns:
\describe{
\item{data}{A data frame}
\item{x}{Any variable you want to visualize}
\item{y}{Another variable you want to visualize}
...
}
}

An object of class \code{data.frame} with 100 rows and 3 columns.
}
\usage{
example_df

example_df
}
\description{
Example data set
}
\examples{
example_df = data.frame(x = rnorm(100), y = rnorm(100),
gender = c(rep('F',35), rep('M',65)))
summarize(example_df, var_list = c(x,y), data_type = 'continuous')
summarize(example_df, var_list = gender, data_type = 'categorical')


}
\keyword{datasets}
