% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.r
\name{graph}
\alias{graph}
\title{Visualize data}
\usage{
graph(
  data = data,
  x = NULL,
  y = NULL,
  cat_var = NULL,
  title = NULL,
  xlab = NULL,
  ylab = NULL,
  fill = NULL
)
}
\arguments{
\item{data}{A data frame containing variables}

\item{x}{Any variable you want to visualize}

\item{y}{Another variable you want to visualize}

\item{cat_var}{Any character variable}

\item{title}{Title of the plot}

\item{xlab}{Label for x axis}

\item{ylab}{Label for y axis}

\item{fill}{Label for the legend}
}
\value{
a graph
}
\description{
Visualize data
}
\examples{

example_df = data.frame(x = rnorm(100), y = rnorm(100),gender = c(rep('F',35), rep('M',65)))
graph(data = example_df, x = example_df$x)

}
