% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.r
\name{summarize}
\alias{summarize}
\title{Clinical Reporting
This package is useful for generating summary statistics for continuous
variables and frequency counts for categorical variables.}
\usage{
summarize(data, var_list, data_type = c("continuous", "categorical"))
}
\arguments{
\item{data}{this is the name of the incoming data set}

\item{var_list}{this is the list of variables}

\item{data_type}{this specifies whether the variables are continuous or
categorical.}
}
\value{
A data frame if var_list contains continuous variables otherwise
it returns a tabyl data frame.
}
\description{
Clinical Reporting
This package is useful for generating summary statistics for continuous
variables and frequency counts for categorical variables.
}
