% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze inflammation.R
\name{analyze}
\alias{analyze}
\title{analyze inflammation}
\usage{
analyze(filename)
}
\arguments{
\item{filename}{The filename of inflammation data in csv format.}
}
\value{

}
\description{
This function analyze mean, maximum, and minimum of inflammation's daily record.
}
\examples{
analyze("data/inflammation-01.csv")
}
